♻️ [youtube/instagram] Split component files for packaging (#32453)
* Split amp-youtube component files

* Split amp-instagram component files

* Move `init` back to AMP element

* Fix test file imports

* Rename allowlisted youtube dependencies on amp-video
