+++ extensions/amp-instagram/1.0/amp-instagram.js
@@ -14,9 +14,8 @@
  * limitations under the License.
  */
 
+import {BaseElement} from './base-element';
 import {CSS} from '../../../build/amp-instagram-1.0.css';
-import {Instagram} from './instagram';
-import {PreactBaseElement} from '../../../src/preact/base-element';
 import {dict} from '../../../src/utils/object';
 import {isExperimentOn} from '../../../src/experiments';
 import {userAssert} from '../../../src/log';
@@ -24,7 +23,7 @@ import {userAssert} from '../../../src/log';
 /** @const {string} */
 const TAG = 'amp-instagram';
 
-class AmpInstagram extends PreactBaseElement {
+class AmpInstagram extends BaseElement {
   /** @override */
   isLayoutSupported(layout) {
     userAssert(
@@ -44,25 +43,6 @@ class AmpInstagram extends PreactBaseElement {
     });
   }
 }
-
-/** @override */
-AmpInstagram['Component'] = Instagram;
-
-/** @override */
-AmpInstagram['children'] = {};
-
-/** @override */
-AmpInstagram['loadable'] = true;
-
-/** @override */
-AmpInstagram['props'] = {
-  'shortcode': {attr: 'data-shortcode'},
-  'captioned': {attr: 'data-captioned'},
-  'title': {attr: 'title'},
-};
-
-/** @override */
-AmpInstagram['layoutSizeDefined'] = true;
 
 AMP.extension(TAG, '1.0', (AMP) => {
   AMP.registerElement(TAG, AmpInstagram, CSS);
