+++ extensions/amp-instagram/1.0/storybook/Basic.js
@@ -21,7 +21,7 @@ import {
   AccordionHeader,
   AccordionSection,
 } from '../../../amp-accordion/1.0/accordion';
-import {Instagram} from '../instagram';
+import {Instagram} from '../component';
 import {boolean, number, text, withKnobs} from '@storybook/addon-knobs';
 import {withA11y} from '@storybook/addon-a11y';
