+++ extensions/amp-youtube/1.0/component.js
