+++ extensions/amp-youtube/1.0/test/test-youtube.js
@@ -15,7 +15,7 @@
  */
 
 import * as Preact from '../../../../src/preact';
-import {Youtube} from '../youtube';
+import {Youtube} from '../component';
 import {dispatchCustomEvent} from '../../../../src/dom';
 import {mount} from 'enzyme';
